(bug) - Fix ruby 3.x incompatibility
It was noted that this gem was failing spec tests when running on ruby 3.x and above.
The root cause of this issue seemed to be the absolute_path property found on the Thread::Backtrace::Location class.
Updating this to path allows spec tests to pass on both ruby 2.x and 3.x.

Can confirm that when consumed by other tools, this also works as expected. Open to suggested changes and improvements.